crypto/ingehackStation/requirements.txt
ecdsa==0.16.1
pycrypto==2.6.1
six==1.15.0
